External block verifier (#916)
Refs: https://github.com/FuelLabs/fuel-core/issues/883

It is the implementation of the block verifier. Right now it provides
only one function to verify the block's fields. But in the future, we
can add verification of the seals(it is why it holds a relayer) and
reuse this verifier in the `fuel-core-sync`.

Along with the verifier, I changed the `fuel_core_poa::Config` and
`ChainConfig` to have rules for the block's time of the network and
separate rules for block production.

- Added `NodeRole::Producer` and `NodeRole::Validator`. Based on the
role, we will create a `fuel_core_poa::Config`.
- Removed `Never` from `BlockProduction` from `ChainConfig` because it
is the block production information of the `PoA`.

No unit tests right now.

Co-authored-by: Tom <tomrgowan@gmail.com>
Co-authored-by: Voxelot <brandonkite92@gmail.com>