fix p2p `max_peers_connected_works` test (#1002)
We've been having certain issues with this test especially in the CI,
here's the analysis and the fix.

First problem is the assertion put on the swarm itself instead of our
PeerManager:
```rust
if p2p_service.swarm.connected_peers().count() > peer_limit { panic!() }
```
While generally this should not return more peers than allowed, one of
the reasons we have PeerManager is to actually contain number of peers
connected logic, so moving the assertion to
```rust
if p2p_service.peer_manager().total_peers_connected() > peer_limit { panic!() }
```
should suffice since PeerManager is the one disconnecting surplus peers.

Other, and bigger, issue with the test was the amount of simulated nodes
we used in our example.
Having 100+ nodes running makes the test itself unnecessarily slow.
This became even more evident now that we made our Heartbeat protocol
ask for a disconnect from a node that did not reply within 2 seconds
with a Block Height. So nodes started disconnecting from each other and
the test would never hit its exit point.

Actually there is no need to have that amount of nodes, instead same
assertion can be achieved with a smaller amount, so I have decreased the
amount of nodes in the test.