feat: Contract state and assets merkle data storage (#1008)
Related PRs:
- Closes https://github.com/FuelLabs/fuel-core/issues/626
- Closes https://github.com/FuelLabs/fuel-core/issues/627

This PR calculates SMT roots for a contract's state and balances. Both
states and balances are modified via random-access inserts and deletes,
so roots must be calculated as an SMT on top of a sparse key-value
dataset.

Additionally, this PR modifies the root calculations and related tests
to use the correct empty root, the 32 byte 0, rather than the hash of
the empty string. The 32 byte 0 is the default for SMTs while the hash
of the empty string is the default for BMTs.

---------

Co-authored-by: green <xgreenx9999@gmail.com>