feat: Add `enable-relayer` flag to `fuel-core` CLI (#1279)
Related issues:
- https://github.com/FuelLabs/fuel-core/issues/1236

This PR adds a new CLI flag to enable the Relayer service
`--enable-relayer`, and disable the Relayer service by default.

When supplying the `--enable-relayer` flag, the `--relayer` argument
becomes mandatory, and omitting it is an error. Similarly, providing a
`--relayer` argument without the `--enable-relayer` flag is an error.

Lastly, providing the `--keypair` or `--network` arguments will also
produce an error if the `--enable-p2p` flag is not set.