Use `AtomicView` in all services (#1612)
Related to https://github.com/FuelLabs/fuel-core/issues/1589 and
preparation for https://github.com/FuelLabs/fuel-core/issues/1583.

All services that use the database for read-only purposes use
`AtomicView` now instead of direct access to the database.

- Removed `Relayer` from the `Verifier` because if it is not used for
now, plus it may not be needed because of the shared sequencer and its
consensus rules.
- Added verification of the transactions root hash into `Verifier`.
- Removed requesting of the one block from p2p, it is possible to use
range request for that purposes.
- Removed not used `get_sealed_header` and `get_sealed_block` method.
- Added the `latest_height` method to `AtomicView` because the database
always knows its latest height.
- Added customisation of the `Height` used by the `AtomicView`. In the
case of the relayer, we want to use `DaBlockHeight` as a primary key to
create a snapshot, while in the case of the Fuel's databases, we want to
use `BockHeight` as a primary key.
- Renamed `Executor` into `ExecutionInstance` and changed it to be a
one-time usable instance. It consumes the `self` to perform the
execution of the block. The instance has a view of the database and
execution options.