Notify services about importing of the genesis block (#1633)
Part of the https://github.com/FuelLabs/fuel-core/issues/1583.

The change moves the genesis block execution and commitment from the
`FuelService::new` to the `FuelService::Task::into_task`.

It allows us to notify other services about the genesis block because
all services are already subscribed to the block importer(it is what we
need for https://github.com/FuelLabs/fuel-core/issues/1583 to process
new messages inside the off-chain worker). Plus, it adds support for the
`async` syntax(it will be used by the parallel regenesis process from
https://github.com/FuelLabs/fuel-core/pull/1519).

Moving genesis block initialization from the constructor to the starting
level breaks p2p because `P2PService` requires knowing the `Genesis`
type to create `FuelP2PService`(It is used to filter connections with
peers). Because of that, I moved the creation of the `FuelP2PService` to
`UninitializedTask::into_task` where the genesis block is already
available.